+++ examples/transaction.rs
@@ -33,6 +33,7 @@ async fn main() {
         .unwrap();
 
     let transaction = client
+        .transaction
         .initialize_transaction(body)
         .await
         .expect("Unable to create transaction");
