+++ src/resources/transaction.rs
@@ -2,8 +2,13 @@
 //! =============
 //! TThe Transactions API allows you create and manage payments on your integration.
 
-use crate::{Channel, Currency};
+use crate::{
+    get_request, post_request, Channel, ChargeBody, Currency, Error, ExportTransactionResponse,
+    PaystackResult, Status, TransactionResponse, TransactionStatus, TransactionStatusList,
+    TransactionTimeline, TransactionTotalsResponse,
+};
 use derive_builder::Builder;
+use reqwest::StatusCode;
 use serde::Serialize;
 
 /// This struct is used to create a transaction body for creating a transaction using the Paystack API.
@@ -71,4 +76,272 @@ pub struct PartialDebitTransactionBody {
     /// Minimum amount to charge
     #[builder(default = "None")]
     at_least: Option<String>,
+}
+
+/// A Struct to hold all the functions of the transaction API route
+#[derive(Debug, Clone)]
+pub struct Transaction {
+    /// Paystack API Key
+    pub api_key: String,
+}
+static BASE_URL: &str = "https://api.paystack.co";
+impl Transaction {
+    /// This method initializes a new transaction using the Paystack API.
+    ///
+    /// It takes a Transaction type as its parameter
+    pub async fn initialize_transaction(
+        &self,
+        transaction_body: InitializeTransactionBody,
+    ) -> PaystackResult<TransactionResponse> {
+        let url = format!("{}/transaction/initialize", BASE_URL);
+
+        match post_request(&self.api_key, &url, transaction_body).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionResponse>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// This method confirms the status of a transaction.
+    ///
+    /// It takes the following parameters:
+    ///     - reference: The transaction reference used to initiate the transaction
+    pub async fn verify_transaction(
+        &self,
+        reference: &String,
+    ) -> PaystackResult<TransactionStatus> {
+        let url = format!("{}/transaction/verify/{}", BASE_URL, reference);
+
+        match get_request(&self.api_key, &url, None).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionStatus>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// This method returns a Vec of transactions carried out on your integrations.
+    ///
+    /// The method takes the following parameters:
+    ///     - perPage (Optional): Number of transactions to return. If None is passed as the parameter, the last 10 transactions are returned.
+    ///     - status (Optional): Filter transactions by status, defaults to Success if no status is passed.
+    ///
+    pub async fn list_transactions(
+        &self,
+        number_of_transactions: Option<u32>,
+        status: Option<Status>,
+    ) -> PaystackResult<TransactionStatusList> {
+        let url = format!("{}/transaction", BASE_URL);
+        let query = vec![
+            ("perPage", number_of_transactions.unwrap_or(10).to_string()),
+            ("status", status.unwrap_or(Status::Success).to_string()),
+        ];
+
+        match get_request(&self.api_key, &url, Some(query)).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionStatusList>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// Get details of a transaction carried out on your integration
+    ///
+    /// This methods take the Id of the desired transaction as a parameter
+    pub async fn fetch_transactions(
+        &self,
+        transaction_id: u32,
+    ) -> PaystackResult<TransactionStatus> {
+        let url = format!("{}/transaction/{}", BASE_URL, transaction_id);
+
+        match get_request(&self.api_key, &url, None).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionStatus>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// All authorizations marked as reusable can be charged with this endpoint whenever you need to receive payments
+    ///
+    /// This function takes a Charge Struct as parameter
+    pub async fn charge_authorization(
+        &self,
+        charge: ChargeBody,
+    ) -> PaystackResult<TransactionStatus> {
+        let url = format!("{}/transaction/charge_authorization", BASE_URL);
+
+        match post_request(&self.api_key, &url, charge).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionStatus>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Charge(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// View the timeline of a transaction
+    ///
+    /// This method takes in the Transaction id or reference as a parameter
+    pub async fn view_transaction_timeline(
+        &self,
+        id: Option<u32>,
+        reference: Option<String>,
+    ) -> PaystackResult<TransactionTimeline> {
+        // This is a hacky implementation to ensure that the transaction reference or id is not empty.
+        // If they are empty, a url without them as parameter is created.
+        let url: PaystackResult<String> = match (id, reference) {
+            (Some(id), None) => Ok(format!("{}/transaction/timeline/{}", BASE_URL, id)),
+            (None, Some(reference)) => {
+                Ok(format!("{}/transaction/timeline/{}", BASE_URL, &reference))
+            }
+            _ => {
+                return Err(Error::Transaction(
+                    "Transaction Id or Reference is need to view transaction timeline".to_string(),
+                ))
+            }
+        };
+
+        let url = url.unwrap(); // Send the error back up the function
+
+        match get_request(&self.api_key, &url, None).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionTimeline>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// Total amount received on your account.
+    ///
+    /// This route normally takes a perPage or page query,
+    /// However in this case it is ignored.
+    /// If you need it in your work please open an issue
+    /// and it will be implemented.
+    pub async fn total_transactions(&self) -> PaystackResult<TransactionTotalsResponse> {
+        let url = format!("{}/transaction/totals", BASE_URL);
+
+        match get_request(&self.api_key, &url, None).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionTotalsResponse>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// Export a list of transactions carried out on your integration
+    ///
+    /// This method takes the following parameters
+    /// - Status (Optional): The status of the transactions to export. Defaults to all
+    /// - Currency (Optional): The currency of the transactions to export. Defaults to NGN
+    /// - Settled (Optional): To state of the transactions to export. Defaults to False.
+    pub async fn export_transaction(
+        &self,
+        status: Option<Status>,
+        currency: Option<Currency>,
+        settled: Option<bool>,
+    ) -> PaystackResult<ExportTransactionResponse> {
+        let url = format!("{}/transaction/export", BASE_URL);
+
+        // Specify a default option for settled transactions.
+        let settled = match settled {
+            Some(settled) => settled.to_string(),
+            None => String::from(""),
+        };
+
+        let query = vec![
+            ("status", status.unwrap_or(Status::Success).to_string()),
+            ("currency", currency.unwrap_or(Currency::EMPTY).to_string()),
+            ("settled", settled),
+        ];
+
+        match get_request(&self.api_key, &url, Some(query)).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<ExportTransactionResponse>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
+
+    /// Retrieve part of a payment from a customer.
+    ///
+    /// It takes a PartialDebitTransaction type as a parameter.
+    ///
+    /// NB: it must be created with the PartialDebitTransaction Builder.
+    pub async fn partial_debit(
+        &self,
+        transaction_body: PartialDebitTransactionBody,
+    ) -> PaystackResult<TransactionStatus> {
+        let url = format!("{}/transaction/partial_debit", BASE_URL);
+
+        match post_request(&self.api_key, &url, transaction_body).await {
+            Ok(response) => match response.status() {
+                StatusCode::OK => match response.json::<TransactionStatus>().await {
+                    Ok(content) => Ok(content),
+                    Err(err) => Err(Error::Transaction(err.to_string())),
+                },
+                _ => Err(Error::RequestNotSuccessful(
+                    response.status().to_string(),
+                    response.text().await?,
+                )),
+            },
+            Err(err) => Err(Error::FailedRequest(err.to_string())),
+        }
+    }
 }
